Use getimagesizefromstring() instead of getimagesize()
getimagesize() outputs warnings when the file does not exist
or when allow_url_fopen is disabled.
Now we use MWHttpRequest to retrieve the file instead
so we don't have a warning if the file can't be retrieved.

Bug: T153250
Change-Id: If1365924e52625a0c6346411faae300b2f83c25a
(cherry picked from commit 676b75aacea1d8d43f0a6c295737033c9b970632)